@@ -16,11 +16,11 @@
         }
     ],
     "require": {
-        "php": ">=5.3.9",
-        "symfony/filesystem": "~2.3|~3.0.0"
+        "php": ">=5.5.9",
+        "symfony/filesystem": "~2.8|~3.0"
     },
     "require-dev": {
-        "symfony/yaml": "~2.7|~3.0.0"
+        "symfony/yaml": "~3.0"
     },
     "suggest": {
         "symfony/yaml": "To use the yaml reference dumper"
@@ -34,7 +34,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "2.8-dev"
+            "dev-master": "3.2-dev"
         }
     }
 }
