@@ -2,10 +2,14 @@
 
 namespace Drupal\better_exposed_filters\Plugin\views\exposed_form;
 
-use Drupal\views\Plugin\views\exposed_form\ExposedFormPluginBase;
+use Drupal\Component\Utility\Html;
 use Drupal\Core\Form\FormStateInterface;
-use Drupal\Core\Url;
+use Drupal\Core\Render\Element;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
+use Drupal\Core\Url;
+use Drupal\views\Plugin\views\exposed_form\InputRequired;
+use Drupal\views\Plugin\views\filter\NumericFilter;
+use Drupal\views\Plugin\views\filter\StringFilter;
 
 /**
  * Exposed form plugin that provides a basic exposed form.
@@ -18,30 +22,7 @@ use Drupal\Core\StringTranslation\StringTranslationTrait;
  *   help = @Translation("Provides additional options for exposed form elements.")
  * )
  */
-class BetterExposedFilters extends ExposedFormPluginBase {
-
-  /**
-   * @inheritdoc
-   */
-  protected function defineOptions() {
-    $options = parent::defineOptions();
-
-    $options['text_input_required'] = array('default' => 'Select any filter and click on Apply to see results', 'translatable' => TRUE);
-    $options['text_input_required_format'] = array('default' => NULL);
-    return $options;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function preExecute() {
-    // Grab BEF settings.
-    $settings = $this->bef_get_settings();
-
-    if (!empty($settings['general']['autosubmit'])) {
-      $this->view->setAjaxEnabled(TRUE);
-    }
-  }
+class BetterExposedFilters extends InputRequired {
 
   /**
    * @inheritdoc
@@ -52,7 +33,26 @@ class BetterExposedFilters extends ExposedFormPluginBase {
     $bef_options = array();
 
     // Get current settings and default values for new filters/
-    $existing = $this->bef_get_settings();
+    $existing = $this->getSettings();
+
+    // Insert a checkbox to make the input required optional just before the
+    // input required text field. Only show the text field if the input required
+    // option is selected.
+    $textInputRequired = $form['text_input_required'];
+    unset($form['text_input_required']);
+    $form['input_required'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Input required'),
+      '#description' => $this->t('Only display results after the user has selected a filter option.'),
+      '#default_value' => !empty($this->options['input_required']),
+    ];
+    $form['text_input_required'] = $textInputRequired += [
+      '#states' => [
+        'visible' => [
+          'input[name="exposed_form_options[input_required]"]' => ['checked' => TRUE],
+        ],
+      ],
+    ];
 
     /*
      * Add general options for exposed form items.
@@ -148,11 +148,18 @@ class BetterExposedFilters extends ExposedFormPluginBase {
           ),
         ),
       );
+
+      // We can only combine sort order and sort by if both options are exposed.
       $bef_options['sort']['advanced']['combine'] = array(
         '#type' => 'checkbox',
         '#title' => $this->t('Combine sort order with sort by'),
         '#default_value' => $existing['sort']['advanced']['combine'],
-        '#description' => $this->t('Combines the sort by options and order (ascending or decending) into a single list.  Use this to display "Option1 (ascending)", "Option1 (descending)", "Option2 (ascending)", "Option2 (descending)" in a single form element.'),
+        '#description' => $this->t('Combines the sort by options and order (ascending or decending) into a single list.  Use this to display "Option1 (ascending)", "Option1 (descending)", "Option2 (ascending)", "Option2 (descending)" in a single form element. Sort order should first be exposed by selecting <em>Allow people to choose the sort order</em>.'),
+        '#states' => [
+          'enabled' => [
+            ':input[name="exposed_form_options[expose_sort_order]"]' => ['checked' => TRUE],
+          ],
+        ],
       );
       $bef_options['sort']['advanced']['combine_rewrite'] = array(
         '#type' => 'textarea',
@@ -162,7 +169,7 @@ class BetterExposedFilters extends ExposedFormPluginBase {
 Post date Asc|Oldest first
 Post date Desc|Newest first
 Title Asc|A -> Z
-Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option altogether.'),
+Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option altogether. The order the options appear will be changed to match the order of options in this field.'),
         '#states' => array(
           'visible' => array(
             ':input[name="exposed_form_options[bef][sort][advanced][combine]"]' => array('checked' => TRUE),
@@ -245,11 +252,14 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
       if (!$bef_filter_intro) {
         $bef_options['bef_intro'] = array(
           '#markup' => '<h3>'
-          . $this->t('Exposed Filter Settings')
-          . '</h3><p>'
-          . $this->t('This section lets you select additional options for exposed filters. Some options are only available in certain situations. If you do not see the options you expect, please see the <a href=":link">BEF settings documentation page</a> for more details.',
-              array(':link' => Url::fromUri('http://drupal.org/node/1701012')->toString()))
-          . '</p>',
+            . $this->t('Exposed Filter Settings')
+            . '</h3><p>'
+            . $this->t('This section lets you select additional options for exposed filters. Some options are only available in certain situations. If you do not see the options you expect, please see the <a href=":link">BEF settings documentation page</a> for more details.',
+              array(
+                ':link' => Url::fromUri('http://drupal.org/node/1701012')
+                  ->toString()
+              ))
+            . '</p>',
         );
         $bef_filter_intro = TRUE;
       }
@@ -313,7 +323,7 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
 
       // The date filter handler extends the numeric filter handler so we have
       // to exclude it specifically.
-      if (is_a($filter, 'Drupal\views\Plugin\views\filter\Numeric') && !is_a($filter, 'Drupal\views\Plugin\views\filter\Date')) {
+      if (is_a($filter, 'Drupal\views\Plugin\views\filter\NumericFilter') && !is_a($filter, 'Drupal\views\Plugin\views\filter\Date')) {
         $bef_slider = TRUE;
       }
 
@@ -350,6 +360,9 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
         $display_options['bef_hidden'] = $this->t('Hidden');
       }
 
+      // Alter the list of available display options for this filter.
+      \Drupal::moduleHandler()->alter('better_exposed_filters_display_options', $display_options, $filter);
+
       $identifier = '"' . $filter->options['expose']['identifier'] . '"';
       if (!empty($filter->options['expose']['label'])) {
         $identifier .= $this->t(' (Filter label: "@fl")', array('@fl' => $filter->options['expose']['label']));
@@ -385,7 +398,6 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
             ),
           ),
           '#description' => $this->t('The minimum allowed value for the jQuery range slider. It can be positive, negative, or zero and have up to 11 decimal places.'),
-          '#element_validate' => array('element_validate_number', 'better_exposed_filters_element_validate_slider_required', 'better_exposed_filters_element_validate_slider_min_max'),
         );
         $bef_options[$label]['slider_options']['bef_slider_max'] = array(
           '#type' => 'textfield',
@@ -398,7 +410,6 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
             ),
           ),
           '#description' => $this->t('The maximum allowed value for the jQuery range slider. It can be positive, negative, or zero and have up to 11 decimal places.'),
-          '#element_validate' => array('element_validate_number', 'better_exposed_filters_element_validate_slider_required', 'better_exposed_filters_element_validate_slider_min_max'),
         );
         $bef_options[$label]['slider_options']['bef_slider_step'] = array(
           '#type' => 'textfield',
@@ -411,9 +422,8 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
             ),
           ),
           '#description' => $this->t('Determines the size or amount of each interval or step the slider takes between the min and max.') . '<br />' .
-              $this->t('The full specified value range of the slider (Range maximum - Range minimum) must be evenly divisible by the step.') . '<br />' .
-              $this->t('The step must be a positive number of up to 5 decimal places.'),
-          '#element_validate' => array('element_validate_number', 'better_exposed_filters_element_validate_slider_required', 'better_exposed_filters_element_validate_slider_step'),
+            $this->t('The full specified value range of the slider (Range maximum - Range minimum) must be evenly divisible by the step.') . '<br />' .
+            $this->t('The step must be a positive number of up to 5 decimal places.'),
         );
         $bef_options[$label]['slider_options']['bef_slider_animate'] = array(
           '#type' => 'textfield',
@@ -421,7 +431,6 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
           '#default_value' => $this->options['bef'][$label]['slider_options']['bef_slider_animate'],
           '#bef_filter_id' => $label,
           '#description' => $this->t('Whether to slide handle smoothly when user click outside handle on the bar. Allowed values are "slow", "normal", "fast" or the number of milliseconds to run the animation (e.g. 1000). If left blank, there will be no animation, the slider will just jump to the new value instantly.'),
-          '#element_validate' => array('better_exposed_filters_element_validate_slider_animate'),
         );
         $bef_options[$label]['slider_options']['bef_slider_orientation'] = array(
           '#type' => 'select',
@@ -500,31 +509,95 @@ Title Desc|Z -> A</pre> Leave the replacement text blank to remove an option alt
         '#description' => $this->t('Places this element in the secondary options portion of the exposed form.'),
       );
 
-      // Allow rewriting of filter options for any filter.
-      $bef_options[$label]['more_options']['rewrite'] = array(
-        '#title' => $this->t('Rewrite filter options'),
-        '#type' => 'details',
-      );
-      $bef_options[$label]['more_options']['rewrite']['filter_rewrite_values'] = array(
-        '#type' => 'textarea',
-        '#title' => $this->t('Rewrite the text displayed'),
-        '#default_value' => $existing[$label]['more_options']['rewrite']['filter_rewrite_values'],
-        '#description' => $this->t('Use this field to rewrite the filter options displayed. Use the format of current_text|replacement_text, one replacement per line. For example: <pre>
-Current|Replacement
-On|Yes
-Off|No
-</pre> Leave the replacement text blank to remove an option altogether. If using hierarchical taxonomy filters, do not including leading hyphens in the current text.
-        '),
-      );
+      if ($filter instanceof StringFilter) {
+        // Allow users to specify placeholder text.
+        $bef_options[$label]['more_options']['placeholder_text'] = [
+          '#type' => 'textfield',
+          '#title' => $this->t('Placeholder text'),
+          '#description' => $this->t('Text to be shown in the text field until it is edited. Leave blank for no placeholder to be set.'),
+          '#default_value' => $existing[$label]['more_options']['placeholder_text'],
+        ];
+      }
+
+      // Allow rewriting of filter options for any filter. String and numeric
+      // filters allow unlimited filter options via textfields, so we can't
+      // offer rewriting for those.
+      // @TODO: check other core filter types to see if there are others that
+      // should be added to this list.
+      if (!$filter instanceof StringFilter && !$filter instanceof NumericFilter) {
+        $bef_options[$label]['more_options']['rewrite'] = array(
+          '#title' => $this->t('Rewrite filter options'),
+          '#type' => 'details',
+        );
+        $bef_options[$label]['more_options']['rewrite']['filter_rewrite_values'] = array(
+          '#type' => 'textarea',
+          '#title' => $this->t('Rewrite the text displayed'),
+          '#default_value' => $existing[$label]['more_options']['rewrite']['filter_rewrite_values'],
+          '#description' => $this->t('Use this field to rewrite the filter options displayed. Use the format of current_text|replacement_text, one replacement per line. For example: <pre>
+  Current|Replacement
+  On|Yes
+  Off|No
+  </pre> Leave the replacement text blank to remove an option altogether. If using hierarchical taxonomy filters, do not including leading hyphens in the current text.
+          '),
+        );
+      }
     }
     /* Ends: foreach ($filters as $filter) { */
-
-    // Alter the list of available display options for this filter.
-    \Drupal::moduleHandler()->alter('better_exposed_filters_display_options', $display_options, $filter);
 
     // Add BEF form elements to the exposed form options form.
     $form['bef'] = $bef_options;
   }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function validateOptionsForm(&$form, FormStateInterface $form_state) {
+    $bef = $form_state->getValue(['exposed_form_options', 'bef']);
+
+    // Validate slider settings.
+    foreach ($bef as $id => $settings) {
+      if ($id == 'general' || $settings['bef_format'] != 'bef_slider') {
+        continue;
+      }
+
+      // Max must be > min.
+      $min = $settings['slider_options']['bef_slider_min'];
+      $max = $settings['slider_options']['bef_slider_max'];
+      if ($max <= $min) {
+        $form_state->setError($form['bef'][$id], $this->t('The slider max value must be greater than the slider min value.'));
+      }
+
+      // Validate the slider animation settings.
+      $animate = $settings['slider_options']['bef_slider_animate'];
+      if ($animate != '') {
+        if ((!is_numeric($animate) || intval($animate) != $animate) && !in_array($animate, ['slow', 'normal', 'fast'])) {
+          $form_state->setError($form['bef'][$id], $this->t('The slider animation option for %name must be "slow", "normal", or "fast", the number of milliseconds to run the animation (e.g. 1000), or blank.', ['%name' => $id]));
+        }
+      }
+
+      // Validate the step setting:
+      //   - Must be positive
+      //   - No more than 5 decimal places
+      //   - Slider range must be evenly divisible by step
+      $step = $settings['slider_options']['bef_slider_step'];
+      if ($step <= 0) {
+        $form_state->setError($form['bef'][$id], $this->t('The slider step option for %name must be a positive number.', ['%name' => $id]));
+      }
+      if (strlen(substr(strrchr((string) $step, '.'), 1)) > 5) {
+        $form_state->setError($form['bef'][$id], $this->t('The slider step option for %name cannot have more than 5 decimal places.', ['%name' => $id]));
+      }
+      // Very small step and a vary large range can go beyond the max value of
+      // an int in PHP. Thus we look for a decimal point when casting the result
+      // to a string.
+      if (strpos((string) ($max - $min) / $step, '.')) {
+        $form_state->setError($form['bef'][$id], $this->t('The slider range for %name must be evenly divisible by the step option.', ['%name' => $id]));
+      }
+    }
+
+    parent::validateOptionsForm($form, $form_state);
+  }
+
+
 
   /**
    * @inheritdoc
@@ -550,11 +623,10 @@ Off|No
       'settings' => array(),
     );
 
-    // Some widgets will require additional CSS.
-    $bef_add_css = FALSE;
-
-    // Grab BEF settings.
-    $settings = $this->bef_get_settings();
+    // Grab BEF settings and allow modules/theme to modify them before
+    // processing.
+    $settings = $this->getSettings();
+    \Drupal::moduleHandler()->alter('better_exposed_filters_settings', $settings, $this->view, $this->displayHandler);
 
     // Some elements may be placed in a secondary details element (eg: "Advanced
     // search options"). Place this after the exposed filters and before the
@@ -563,7 +635,6 @@ Off|No
       $secondary = array(
         '#type' => 'details',
         '#title' => $settings['general']['secondary_label'],
-        '#theme' => 'secondary_exposed_elements',
       );
     }
 
@@ -571,7 +642,7 @@ Off|No
     if (!empty($settings['general']['autosubmit'])) {
       $form = array_merge_recursive($form, array('#attributes' => array('data-bef-auto-submit-full-form' => '')));
       $form['actions']['submit']['#attributes']['data-bef-auto-submit-click'] = '';
-      $form['#attached']['library'] = ['better_exposed_filters/auto_submit'];
+      $form['#attached']['library'][] = 'better_exposed_filters/auto_submit';
 
       if (!empty($settings['general']['autosubmit_hide'])) {
         $form['actions']['submit']['#attributes']['class'][] = 'js-hide';
@@ -581,7 +652,7 @@ Off|No
     /*
      * Handle exposed sort elements.
      */
-    if (isset($settings['sort']) && !empty($form['sort_by']) && !empty($form['sort_order'])) {
+    if (isset($settings['sort']) && !empty($form['sort_by'])) {
       $show_apply = TRUE;
 
       // If selected, collect all sort-related form elements and put them
@@ -591,7 +662,7 @@ Off|No
       $sort_elems = array();
 
       // Check for combined sort_by and sort_order.
-      if ($settings['sort']['advanced']['combine']) {
+      if ($settings['sort']['advanced']['combine'] && !empty($form['sort_order'])) {
         $options = [];
 
         // Add reset sort option at the top of the list.
@@ -619,7 +690,7 @@ Off|No
 
         // Rewrite the option values if any were specified.
         if (!empty($settings['sort']['advanced']['combine_rewrite'])) {
-          $options = $this->rewriteOptions($options, $settings['sort']['advanced']['combine_rewrite']);
+          $options = $this->rewriteOptions($options, $settings['sort']['advanced']['combine_rewrite'], TRUE);
           if (!isset($options[$selected])) {
             // Avoid "illegal choice" errors if the selected option is
             // eliminated by the rewrite.
@@ -686,19 +757,25 @@ Off|No
         }
         elseif ('bef_links' == $settings['sort']['bef_format']) {
           $form['sort_by']['#theme'] = 'bef_links';
-          $form['sort_order']['#theme'] = 'bef_links';
+          if(!empty($form['sort_order'])) {
+            $form['sort_order']['#theme'] = 'bef_links';
+          }
 
           // Exposed form displayed as blocks can appear on pages other than the
           // view results appear on. This can cause problems with
           // select_as_links options as they will use the wrong path. We provide
           // a hint for theme functions to correct this.
           $form['sort_by']['#bef_path'] = $this->displayHandler->getUrl();
-          $form['sort_order']['#bef_path'] = $this->displayHandler->getUrl();
+          if(!empty($form['sort_order'])) {
+            $form['sort_order']['#bef_path'] = $this->displayHandler->getUrl();
+          }
         }
 
         if ($collapse) {
           $sort_elems[] = 'sort_by';
-          $sort_elems[] = 'sort_order';
+          if(!empty($form['sort_order'])) {
+            $sort_elems[] = 'sort_order';
+          }
         }
 
         // Add reset sort option if selected.
@@ -788,10 +865,16 @@ Off|No
       // configurable.
       $field_id = $filters[$label]->options['expose']['identifier'];
 
+      // Check for placeholder text.
+      if (!empty($settings[$label]['more_options']['placeholder_text'])) {
+        // @todo: Add token replacement for placeholder text.
+        $form[$label]['#placeholder'] = $settings[$label]['more_options']['placeholder_text'];
+      }
+
       // Handle filter value rewrites.
       if (!empty($options['more_options']['rewrite']['filter_rewrite_values'])) {
         $form[$field_id]['#options'] = $this->rewriteOptions($form[$field_id]['#options'] , $options['more_options']['rewrite']['filter_rewrite_values']);
-        if (!isset($form[$field_id]['#options'][$selected])) {
+        if (isset($selected) && !isset($form[$field_id]['#options'][$selected])) {
           // Avoid "Illegal choice" errors.
           $form[$field_id]['#default_value'] = NULL;
         }
@@ -971,20 +1054,19 @@ Off|No
 
         case 'bef_slider':
           $show_apply = TRUE;
-          $bef_add_js = TRUE;
-          $bef_add_css = TRUE;
-          $bef_js['slider'] = TRUE;
+          $form[$field_id]['#attached']['library'][] = 'core/jquery.ui.slider';
+          $form[$field_id]['#attached']['library'][] = 'better_exposed_filters/sliders';
 
-          // Add js options for the slider for this filter.
-          $bef_js['slider_options'][$field_id] = array(
+          $form[$field_id]['#attached']['drupalSettings']['better_exposed_filters']['slider'] = TRUE;
+            $form[$field_id]['#attached']['drupalSettings']['better_exposed_filters']['slider_options'][$field_id] = [
             'min' => $options['slider_options']['bef_slider_min'],
             'max' => $options['slider_options']['bef_slider_max'],
             'step' => $options['slider_options']['bef_slider_step'],
             'animate' => $options['slider_options']['bef_slider_animate'],
             'orientation' => $options['slider_options']['bef_slider_orientation'],
-            'id' => drupal_html_id($field_id),
+            'id' => Html::getUniqueId($field_id),
             'viewId' => $form['#id'],
-          );
+          ];
           break;
 
         case 'bef_links':
@@ -1010,9 +1092,6 @@ Off|No
           $form[$field_id]['#title'] = $filters[$label]->options['expose']['label'];
           $form[$field_id]['#return_value'] = 1;
           $form[$field_id]['#type'] = 'checkbox';
-
-          // Handoff to the theme layer.
-          $form[$field_id]['#theme'] = 'checkbox';
           break;
 
         case 'bef':
@@ -1025,25 +1104,6 @@ Off|No
           // Clean up objects from the options array (happens for taxonomy-
           // based filters).
           $form[$field_id]['#options'] = $this->cleanOptions($form[$field_id]['#options']);
-
-          if (!empty($options['more_options']['bef_collapsible'])) {
-            // Pass the description and title along in a way such that it does
-            // not get rendered as part of the exposed form widget.  We render
-            // them as part of the details element instead.
-            $form[$field_id]['#theme_wrappers'] = [
-              'details' => [
-                '#title' => $form['#info']["filter-$label"]['label'] ?: '',
-                '#description' => $form['#info']["filter-$label"]['description'] ?: '',
-                // Needed to keep styling consistent with other exposed options.
-                '#attributes' => array('class' => 'form-item'),
-              ],
-            ];
-            $form['#info']["filter-$label"]['label'] = '';
-            $form['#info']["filter-$label"]['description'] = '';
-
-            // @TODO: Handle filters with exposed operators -- they need to end
-            // up in the same details element.
-          }
 
           // Render as either radio buttons or checkboxes.
           if (empty($form[$field_id]['#multiple'])) {
@@ -1060,7 +1120,7 @@ Off|No
             $form[$field_id]['#theme'] = 'bef_checkboxes';
 
             if ($options['more_options']['bef_select_all_none'] || $options['more_options']['bef_select_all_none_nested']) {
-              $form[$field_id]['#attached']['library'] = ['better_exposed_filters/select_all_none'];
+              $form[$field_id]['#attached']['library'][] = 'better_exposed_filters/select_all_none';
 
               $form[$field_id]['#bef_select_all_none'] = $options['more_options']['bef_select_all_none'];
               $form[$field_id]['#bef_select_all_none_nested'] = $options['more_options']['bef_select_all_none_nested'];
@@ -1089,20 +1149,53 @@ Off|No
       }
       /* Ends switch ($options['bef_format']) */
 
+      // Collapsible filters are an option for all exposed filters.
+      if (!empty($options['more_options']['bef_collapsible'])) {
+        // Pass the description and title along in a way such that it does
+        // not get rendered as part of the exposed form widget.  We render
+        // them as part of the details element instead.
+        $form[$field_id]['#theme_wrappers'] = [
+          'details' => [
+            '#title' => $form['#info']["filter-$label"]['label'] ?: '',
+            '#description' => $form['#info']["filter-$label"]['description'] ?: '',
+            // Needed to keep styling consistent with other exposed options.
+            '#attributes' => array('class' => 'form-item'),
+            '#value' => NULL,
+          ],
+        ];
+        $form['#info']["filter-$label"]['label'] = '';
+        $form['#info']["filter-$label"]['description'] = '';
+
+        // Check if the operator is exposed for this filter.
+        if ($this->view->getHandlers('filter')[$field_id]['expose']['use_operator']) {
+          // Include the exposed operator with the filter.
+          $operator_id = $this->view->getHandlers('filter')[$field_id]['expose']['operator_id'];
+          $form[$field_id][$operator_id] = $form[$operator_id];
+          unset($form[$operator_id]);
+        }
+      }
+
       // Check if this is a secondary form element.
       if ($allow_secondary && $settings[$label]['more_options']['is_secondary']) {
-        $identifier = $form['#info']["filter-$label"]['value'];
+        if ($filters[$label]->options['is_grouped']) {
+          $identifier = $filters[$label]->options['group_info']['identifier'];
+          $filter_info_name = "filter-$identifier";
+        }
+        else {
+          $identifier = $filters[$label]->options['expose']['identifier'];
+          $filter_info_name = "filter-$label";
+        }
         if (!empty($form[$identifier])) {
           // Move exposed operators with exposed filters
-          if (!empty($this->display->display_options['filters'][$identifier]['expose']['use_operator'])) {
-            $op_id = $this->display->display_options['filters'][$identifier]['expose']['operator_id'];
+          if (!empty($this->view->getHandlers('filter')[$label]['expose']['use_operator'])) {
+            $op_id = $this->view->getHandlers('filter')[$label]['expose']['operator_id'];
             $secondary[$op_id] = $form[$op_id];
             unset($form[$op_id]);
           }
           $secondary[$identifier] = $form[$identifier];
           unset($form[$identifier]);
-          $secondary[$identifier]['#title'] = $form['#info']["filter-$label"]['label'];
-          unset($form['#info']["filter-$label"]);
+          $secondary[$identifier]['#title'] = $form['#info'][$filter_info_name]['label'];
+          unset($form['#info'][$filter_info_name]);
         }
       }
     }
@@ -1118,20 +1211,34 @@ Off|No
         $form['#attached']['library'][] = 'core/jquery.ui.datepicker';
         $form['#attached']['library'][] = 'better_exposed_filters/datepickers';
       }
-      if ($bef_js['slider']) {
-        $form['#attached']['library'][] = 'core/jquery.ui.slider';
-      }
 
       $form['#attached']['drupalSettings']['better_exposed_filters'] = $bef_js;
     }
 
     // Check for secondary elements.
     if ($allow_secondary && !empty($secondary)) {
-      // Add secondary elements after regular exposed filter elements.
-      $remaining = array_splice($form, count($form['#info']) + 1);
+      // Use weights to place the secondary elements just ahead of the
+      // submit/reset buttons, but after all the other expsoed elements.
+      foreach(Element::children($form) as $index => $child) {
+        $form[$child]['#weight'] = $index;
+      }
       $form['secondary'] = $secondary;
-      $form = array_merge($form, $remaining);
+      $form['secondary']['#weight'] = count($form['#info']) -1;
       $form['#info']['filter-secondary']['value'] = 'secondary';
+    }
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function exposedFilterApplied() {
+    // If the input required option is set, check to see if a filter option has
+    // been set.
+    if (!empty($this->options['input_required'])) {
+      return parent::exposedFilterApplied();
+    }
+    else {
+      return TRUE;
     }
   }
 
@@ -1145,38 +1252,82 @@ Off|No
    *   An array of key => value pairs that may be rewritten.
    * @param string $rewriteSettings
    *   String representing the entry in the settings form.
+   * @param bool $reorder
+   *   Reorder $options based on the rewrite settings.
    *
    * @return array
    *   Rewritten $options.
    */
-  protected function rewriteOptions(array $options, $rewriteSettings) {
+  protected function rewriteOptions(array $options, $rewriteSettings, $reorder = FALSE) {
     if (empty($rewriteSettings) || !is_string($rewriteSettings)) {
       return $options;
     }
     $rewrites = [];
+    $order = [];
+    if (!$reorder) {
+      $order = array_keys($options);
+    }
     $lines = explode("\n", trim($rewriteSettings));
     foreach ($lines as $line) {
       list($search, $replace) = explode('|', $line);
       if (!empty($search)) {
         $rewrites[$search] = $replace;
+        if ($reorder) {
+          // Reorder options in the order they are specified in rewrites.
+          // Collect the keys to use later.
+          // @TODO: need to handle taxonomy terms before exposing this option.
+          $key = array_search($search, $options);
+          if ($key !== FALSE) {
+            $order[] = $key;
+          }
+        }
       }
     }
 
-    $return = $options;
-    // @TODO: Need to handle rewriting of taxonomy filters.
-    // https://www.drupal.org/node/2666540
-    foreach ($options as $index => $option) {
-      // Some options, such as "- Any -", are passed as TranslatableMarkup
-      // objects. Convert them to strings for the comparison.
-      if (!is_string($option)) {
+    $return = [];
+    if ($reorder && !empty($order)) {
+      // Start with the items that were listed in the rewrite settings.
+      foreach ($order as $key) {
+        $return[$key] = $options[$key];
+        unset($options[$key]);
+      }
+
+      // Tack remaaining options on the end.
+      $return += $options;
+    }
+    else {
+      $return = $options;
+    }
+
+    // Rewrite the option text.
+    foreach ($return as $index => $option) {
+      // Handle taxonomy terms which use a stdClass option object that doesn't
+      // implement __toString().
+      if ($option instanceof \stdClass && isset($option->option) && is_array($option->option)) {
+        $option = array_values($option->option)[0];
+      }
+      else if (!is_string($option)) {
+        // Some options, such as "- Any -", are passed as TranslatableMarkup
+        // objects. Convert them to strings for the comparison.
         $option = (string) $option;
       }
+      if (!is_string($option)) {
+        // We give up...
+        continue;
+      }
+
       if (isset($rewrites[$option])) {
         if ('' == $rewrites[$option]) {
           unset($return[$index]);
         }
         else {
-          $return[$index] = $rewrites[$option];
+          if ($return[$index] instanceof \stdClass) {
+            list($tid, $text) = each($return[$index]->option);
+            $return[$index]->option[$tid] = $rewrites[$text];
+          }
+          else {
+            $return[$index] = $rewrites[$option];
+          }
         }
       }
     }
@@ -1223,12 +1374,12 @@ Off|No
    * values.  Use this to set defaults for missing values in a multi-dimensional
    * array.  Eg:
    *
-   *  $existing = $this->bef_set_defaults($defaults, $existing);
+   *  $existing = $this->setDefaults($defaults, $existing);
    *
    * @return array
    *   The resulting settings array
    */
-  protected function bef_set_defaults() {
+  protected function setDefaults() {
     $count = func_num_args();
     if (!$count) {
       return;
@@ -1247,7 +1398,7 @@ Off|No
         // Numeric keyed values are added (unless already there).
         if (is_numeric($key) && !in_array($value, $return)) {
           if (is_array($value)) {
-            $return[] = $this->bef_set_defaults($return[$key], $value);
+            $return[] = $this->setDefaults($return[$key], $value);
           }
           else {
             $return[] = $value;
@@ -1256,7 +1407,7 @@ Off|No
         // String keyed values are replaced.
         else {
           if (isset($return[$key]) && is_array($value) && is_array($return[$key])) {
-            $return[$key] = $this->bef_set_defaults($return[$key], $value);
+            $return[$key] = $this->setDefaults($return[$key], $value);
           }
           else {
             $return[$key] = $value;
@@ -1273,7 +1424,7 @@ Off|No
    * @param array $settings
    *   Array of BEF settings.
    */
-  protected function bef_update_legacy_settings($settings) {
+  protected function updateLegacySettings($settings) {
     // There has got to be a better way... But for now, this works.
     if (isset($settings['sort']['collapsible'])) {
       $settings['sort']['advanced']['collapsible'] = $settings['sort']['collapsible'];
@@ -1308,12 +1459,12 @@ Off|No
    * @endcode
    * as there will be a default value at all positions in the settings array.
    * Also updates legacy settings to their new locations via
-   * bef_update_legacy_settings().
+   * updateLegacySettings().
    *
    * @return array
    *   Multi-dimensional settings array.
    */
-  protected function bef_get_settings() {
+  protected function getSettings() {
     // General, sort, pagers, etc.
     $defaults = array(
       'general' => array(
@@ -1343,15 +1494,19 @@ Off|No
     // Update legacy settings in the exposed form settings form. This
     // keep us from losing settings when an option is put into an
     // 'advanced options' details element.
-    $current = $this->bef_update_legacy_settings($this->options['bef']);
+    $current = [];
+    if (isset($this->options['bef'])) {
+      $current = $this->updateLegacySettings($this->options['bef']);
+    }
 
     // Collect existing values or use defaults.
-    $settings = $this->bef_set_defaults($defaults, $current);
+    $settings = $this->setDefaults($defaults, $current);
 
     // Filter default values.
     $filter_defaults = array(
       'bef_format' => 'default',
       'more_options' => array(
+        'placeholder_text' => '',
         'bef_select_all_none' => FALSE,
         'bef_select_all_none_nested' => FALSE,
         'bef_collapsible' => FALSE,
@@ -1381,9 +1536,10 @@ Off|No
         $settings[$label] = $filter_defaults;
       }
       else {
-        $settings[$label] = $this->bef_set_defaults($filter_defaults, $this->options['bef'][$label]);
+        $settings[$label] = $this->setDefaults($filter_defaults, $this->options['bef'][$label]);
       }
     }
     return $settings;
   }
+
 }
