@@ -81,7 +81,7 @@ class Drupal {
   /**
    * The current system version.
    */
-  const VERSION = '8.3.5';
+  const VERSION = '8.3.6';
 
   /**
    * Core API compatibility.
