@@ -275,7 +275,7 @@ class ViewEditForm extends ViewFormBase {
 
     // Rename display ids if needed.
     foreach ($executable->displayHandlers as $id => $display) {
-      if (!empty($display->display['new_id']) && empty($display->display['deleted'])) {
+      if (!empty($display->display['new_id']) && $display->display['new_id'] !== $display->display['id'] && empty($display->display['deleted'])) {
         $new_id = $display->display['new_id'];
         $display->display['id'] = $new_id;
         unset($display->display['new_id']);
@@ -289,6 +289,9 @@ class ViewEditForm extends ViewFormBase {
           'view' => $view->id(),
           'display_id' => $new_id,
         ]);
+      }
+      elseif (isset($display->display['new_id'])) {
+        unset($display->display['new_id']);
       }
     }
     $view->set('display', $displays);
