@@ -10,16 +10,16 @@
  *
  * @param $settings
  *   The settings array.
- * @param $context
- *   The view and display to which the settings apply.
- *
- * @todo:
-function hook_better_exposed_filters_settings_alter(&$settings, $context) {
+ * @param $view
+ *   The view to which the settings apply.
+ * @param $displayHandler
+ *   The display handler to which the settings apply.
+ */
+function hook_better_exposed_filters_settings_alter(&$settings, $view, $displayHandler) {
   // Set the min/max value of a slider.
   $settings['field_price_value']['slider_options']['bef_slider_min'] = 500;
   $settings['field_price_value']['slider_options']['bef_slider_max'] = 5000;
 }
- */
 
 
 /**
