@@ -3,7 +3,7 @@
 namespace Drupal\Tests\metatag\Functional;
 
 use Drupal\user\Entity\User;
-
+use Drupal\Component\Utility\Html;
 
 /**
  * Misc helper functions for the automated tests.
@@ -23,6 +23,23 @@ trait MetatagHelperTrait {
     $account->passRaw = $password;
     $account->pass_raw = $password;
     $this->drupalLogin($account);
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  protected function verbose($message, $title = NULL) {
+    // Handle arrays, objects, etc.
+    if (!is_string($message)) {
+      $message = "<pre>\n" . print_r($message, TRUE) . "\n</pre>\n";
+    }
+
+    // Optional title to go before the output.
+    if (!empty($title)) {
+      $title = '<h2>' . Html::escape($title) . "</h2>\n";
+    }
+
+    parent::verbose($title . $message);
   }
 
 }
